Major Update
After a discussion with my supervisor, I'm implementing a number of major changes with the way traders are defined. Traders will now be fluid between flavours and potentially change over time. There is now only one type of technical analyst. The mecahism for enterting the market is now utility-based rather than based on a discrete rule. The utility-based rule now also determines the quantity of the order placed.