Removal of Pymarket
The simulation will no longer use pymarket and I will instead be designing the market mechanism myself. All the code pertaining to pymarket has been moved to another notebook called "old" so it is available if needed. I have also removed the use of RSI and stochastic oscillators in order to simplify the model.